+++ packages/use-schema-org-action/src/test/useSchemaOrgAction.spec.ts
@@ -4,7 +4,7 @@ import { act } from '@testing-library/react';
 import { type PartialDeep } from 'type-fest';
 import { type ActionStatusType } from '../ActionStatusType';
 import { type PropertyValueSpecification } from '../PropertyValueSpecificationSchema';
-import useSchemaOrgAction, { type ActionHandler } from '../useSchemaOrgAction2';
+import useSchemaOrgAction, { type ActionHandler } from '../useSchemaOrgAction';
 import { type MockOf } from './MockOf';
 import renderHook, { type RenderHookResult } from './renderHook';
 import sortEntries from './sortEntries';
